CodeClimate fixes
